+++ src/pyiem/nws/products/spcpts.py
@@ -340,11 +340,11 @@ class SPCPTS(TextProduct):
     def draw_outlooks(self):
         """For debugging, draw the outlooks on a simple map for inspection."""
         # pylint: disable=import-outside-toplevel
-        import matplotlib.pyplot as plt
+        from pyiem.plot.use_agg import figure
 
         for day, collect in self.outlook_collections.items():
             for outlook in collect.outlooks:
-                fig = plt.figure(figsize=(12, 8))
+                fig = figure(figsize=(12, 8))
                 ax = fig.add_subplot(111)
                 # pylint: disable=unsubscriptable-object
                 ax.plot(
@@ -380,7 +380,6 @@ class SPCPTS(TextProduct):
                 ).replace(" ", "_")
                 LOG.warning(":: creating plot %s", fn)
                 fig.savefig(fn)
-                plt.close()
 
     def set_metadata(self):
         """Set some metadata about this product."""