+++ CHANGELOG.md
@@ -7,6 +7,8 @@ All notable changes to this library are documented in this file.
 
 ### API Changes
 
+- Remove `pyiem.plot.use_agg.plt` (pyplot) as it is a walking memory leak,
+
 ### New Features
 
 ### Bug Fixes